+++ src/test/java/com/lge/lai/common/db/dao/AsbMimeDAOTest.java
@@ -30,21 +30,16 @@ public class AsbMimeDAOTest {
         asb = new Asb(version, type, desc, packageName, className, actionName, updatedBy);
 
         AsbDAO asbDAO = daoFactory.getAsbDAO();
-        try {
-            asbId = asbDAO.create(asb);
-        } catch (DAOException e) {
-            fail("creating asb failed: " + e);
+        asbId = asbDAO.create(asb);
+        if (asbId == -1) {
+            fail("duplicate entry exists");
         }
     }
 
     @After
     public void tearDown() {
         AsbDAO asbDAO = daoFactory.getAsbDAO();
-        try {
-            asbDAO.delete(asbId);
-        } catch (DAOException e) {
-            fail("deleting asb failed: " + e);
-        }
+        asbDAO.delete(asbId);
     }
 
     @Test
@@ -54,21 +49,14 @@ public class AsbMimeDAOTest {
                 asb.className, asb.actionName, mimeType, asb.updatedBy);
 
         AsbMimeDAO asbMimeDAO = daoFactory.getAsbMimeDAO(asbId);
-        long asbMimeId = 0L;
-        try {
-            asbMimeId = asbMimeDAO.create(asbMime);
-            AsbMime insertedRow = (AsbMime)asbMimeDAO.find(asbMimeId);
-            if (!insertedRow.equals(asbMime)) {
-                fail("created object is not same with expected object");
-            }
-        } catch (DAOException e) {
-            fail("creating asbMime failed: " + e);
-        } finally {
-            try {
-                asbMimeDAO.delete(asbMimeId);
-            } catch (DAOException e) {
-                fail("deleting asbMime failed: " + e);
-            }
+        long asbMimeId = asbMimeDAO.create(asbMime);
+        if (asbMimeId == -1) {
+            fail("duplicated entry exists");
         }
+        AsbMime insertedRow = (AsbMime)asbMimeDAO.find(asbMimeId);
+        if (!insertedRow.equals(asbMime)) {
+            fail("created object is not same with expected object");
+        }
+        asbMimeDAO.delete(asbMimeId);
     }
 }
